requirements.txt
pyserial
argparse
configparser